// frontend/lib/generated/fastapi/schemas/planType.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */

export type PlanType = typeof PlanType[keyof typeof PlanType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PlanType = {
  one_time_price: 'one_time_price',
  recurring_price: 'recurring_price',
} as const;