// frontend/lib/generated/fastapi/schemas/publicPlan.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicPlanDescription } from './publicPlanDescription';
import type { PublicPlanDiscount } from './publicPlanDiscount';
import type { PublicPlanExternalId } from './publicPlanExternalId';
import type { PublicPlanPaymentFrequencyCount } from './publicPlanPaymentFrequencyCount';
import type { PublicPlanPaymentFrequencyInterval } from './publicPlanPaymentFrequencyInterval';
import type { PublicPlanSubscriptionPeriodCount } from './publicPlanSubscriptionPeriodCount';
import type { PublicPlanSubscriptionPeriodInterval } from './publicPlanSubscriptionPeriodInterval';
import type { PublicPlanFeature } from './publicPlanFeature';

export interface PublicPlan {
  id: string;
  name: string;
  slug: string;
  description?: PublicPlanDescription;
  type: string;
  price: number;
  discount?: PublicPlanDiscount;
  boostForDays?: number;
  isDefault?: boolean;
  externalId?: PublicPlanExternalId;
  paymentFrequencyCount?: PublicPlanPaymentFrequencyCount;
  paymentFrequencyInterval?: PublicPlanPaymentFrequencyInterval;
  subscriptionPeriodCount?: PublicPlanSubscriptionPeriodCount;
  subscriptionPeriodInterval?: PublicPlanSubscriptionPeriodInterval;
  productCount?: number;
  features?: PublicPlanFeature[];
}