// frontend/lib/generated/fastapi/schemas/publicProductCard.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicProductCardCategory } from './publicProductCardCategory';
import type { PublicProductCardCreatedAt } from './publicProductCardCreatedAt';
import type { PublicProductCardUpdatedAt } from './publicProductCardUpdatedAt';
import type { PublicProductCardLatestRevenueCents } from './publicProductCardLatestRevenueCents';
import type { PublicProductCardRevenueCurrencyCode } from './publicProductCardRevenueCurrencyCode';
import type { PublicProductCardScoreCount } from './publicProductCardScoreCount';
import type { PublicProductCardInterest } from './publicProductCardInterest';

export interface PublicProductCard {
  id: string;
  slug: string;
  name: string;
  logo: string;
  tagline: string;
  category?: PublicProductCardCategory;
  badges?: string[];
  sponsored?: boolean;
  isVerified?: boolean;
  createdAt?: PublicProductCardCreatedAt;
  updatedAt?: PublicProductCardUpdatedAt;
  latestRevenueCents?: PublicProductCardLatestRevenueCents;
  revenueCurrencyCode?: PublicProductCardRevenueCurrencyCode;
  scoreCount?: PublicProductCardScoreCount;
  interest?: PublicProductCardInterest;
}