// frontend/lib/generated/fastapi/schemas/publicProductDetailPayload.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicProductDetailPayloadBannerImage } from './publicProductDetailPayloadBannerImage';
import type { PublicProductDetailPayloadPricingModel } from './publicProductDetailPayloadPricingModel';
import type { PublicProductDetailPayloadStartingPriceCents } from './publicProductDetailPayloadStartingPriceCents';
import type { PublicProductDetailPayloadCurrencyCode } from './publicProductDetailPayloadCurrencyCode';
import type { PublicProductDetailPayloadType } from './publicProductDetailPayloadType';
import type { PublicProductDetailPayloadPublishedAt } from './publicProductDetailPayloadPublishedAt';
import type { PublicProductDetailPayloadCreatedAt } from './publicProductDetailPayloadCreatedAt';
import type { PublicProductDetailPayloadUpdatedAt } from './publicProductDetailPayloadUpdatedAt';
import type { PublicProductDetailPayloadCategory } from './publicProductDetailPayloadCategory';
import type { PublicProductAlternative } from './publicProductAlternative';
import type { PublicProductDetailPayloadUser } from './publicProductDetailPayloadUser';
import type { PublicProductDetailPayloadMetadata } from './publicProductDetailPayloadMetadata';
import type { PublicProductDetailPayloadAnalytics } from './publicProductDetailPayloadAnalytics';
import type { PublicProductDetailPayloadVerification } from './publicProductDetailPayloadVerification';
import type { PublicProductDetailPayloadInterest } from './publicProductDetailPayloadInterest';
import type { PublicProductMedia } from './publicProductMedia';
import type { PublicPlanFeatureAssignment } from './publicPlanFeatureAssignment';
import type { PublicProductDetailPayloadUpvotesCount } from './publicProductDetailPayloadUpvotesCount';

export interface PublicProductDetailPayload {
  id: string;
  slug: string;
  name: string;
  tagline: string;
  description: string;
  websiteUrl: string;
  logo: string;
  bannerImage?: PublicProductDetailPayloadBannerImage;
  pricingModel?: PublicProductDetailPayloadPricingModel;
  startingPriceCents?: PublicProductDetailPayloadStartingPriceCents;
  currencyCode?: PublicProductDetailPayloadCurrencyCode;
  platforms?: string[];
  status: string;
  type?: PublicProductDetailPayloadType;
  publishedAt?: PublicProductDetailPayloadPublishedAt;
  createdAt?: PublicProductDetailPayloadCreatedAt;
  updatedAt?: PublicProductDetailPayloadUpdatedAt;
  keywords?: string[];
  category?: PublicProductDetailPayloadCategory;
  alternatives?: PublicProductAlternative[];
  user?: PublicProductDetailPayloadUser;
  metadata?: PublicProductDetailPayloadMetadata;
  analytics?: PublicProductDetailPayloadAnalytics;
  verification?: PublicProductDetailPayloadVerification;
  interest?: PublicProductDetailPayloadInterest;
  media?: PublicProductMedia[];
  badges?: string[];
  planAssignments?: PublicPlanFeatureAssignment[];
  activeFeatureEntitlements?: string[];
  upvotesCount?: PublicProductDetailPayloadUpvotesCount;
}