// frontend/lib/generated/fastapi/schemas/publicProductLeaderboardScore.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicProductLeaderboardScoreRank } from './publicProductLeaderboardScoreRank';

export interface PublicProductLeaderboardScore {
  points: number;
  rank?: PublicProductLeaderboardScoreRank;
  available?: boolean;
}